+++ muacrypt/mime.py
@@ -206,6 +206,7 @@ class ACParseResult(object):
 def gen_mail_msg(From, To, Cc=None, _extra=None, Autocrypt=None,
                  Subject="testmail", Date=None, _dto=False,
                  MessageID=None, payload='Autoresponse\n',
+                 ENCRYPT=None,
                  charset=None):
     if Cc is None:
         Cc = []
@@ -227,6 +228,8 @@ def gen_mail_msg(From, To, Cc=None, _extra=None, Autocrypt=None,
     msg['Message-ID'] = MessageID
     if Subject is not None:
         msg['Subject'] = Subject
+    if ENCRYPT is not None:
+        msg['ENCRYPT'] = ENCRYPT
     Date = 0 if not Date else Date
     if isinstance(Date, int):
         Date = formatdate(time.time() + Date)