+++ test_muacrypt/test_cmdline.py
@@ -180,7 +180,7 @@ class TestProcessOutgoing:
         # mycmd.run_fail(["process-outgoing"], input=mail.as_string(), fnl="""
         #     *AccountNotFound*x@y.org*
         # """)
-        out0 = mycmd.run_ok(["process-outgoing"], input=mail.as_string())
+        out0 = mycmd.run_fail(["process-outgoing"], input=mail.as_string())
         assert "Autocrypt" not in out0
 
         mail = gen_mail(From="account1@a.org")
@@ -221,18 +221,18 @@ class TestProcessOutgoing:
         mycmd.run_ok(["mod-account", "default", "--email-regex", "123123"])
         mail = gen_mail().as_string()
         pargs = ["-oi", "b@b.org"]
-        mycmd.run_ok(["sendmail", "-f", "--"] + pargs, input=mail)
-        assert len(popen_mock.calls) == 1
-        call = popen_mock.pop_next_call()
-        for x in pargs:
-            assert x in call.args
-        # make sure unknown option is passed to pipe
-        assert "-f" in call.args
-        out_msg = mime.parse_message_from_string(call.input)
-        assert "Autocrypt" not in out_msg, out_msg.as_string()
+        mycmd.run_fail(["sendmail", "-f", "--"] + pargs, input=mail)
+        # assert len(popen_mock.calls) == 1
+        # call = popen_mock.pop_next_call()
+        # for x in pargs:
+        #     assert x in call.args
+        # # make sure unknown option is passed to pipe
+        # assert "-f" in call.args
+        # out_msg = mime.parse_message_from_string(call.input)
+        # assert "Autocrypt" not in out_msg, out_msg.as_string()
 
     def test_sendmail_fails(self, mycmd, gen_mail, popen_mock):
-        mycmd.run_ok(["add-account", "default", "--email-regex=account1@a.org"])
+        mycmd.run_ok(["add-account", "default", "--email-regex=.*"])
         mail = gen_mail().as_string()
         pargs = ["-oi", "b@b.org"]
         popen_mock.mock_next_call(ret=2)